% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/He_2019_metadata.R
\docType{data}
\name{He_2019_metadata}
\alias{He_2019_metadata}
\title{He_2019_data Metadata}
\format{
A data frame with 208 rows and 51 columns:
\describe{
  \item{X}{Description of X}
  \item{Subject}{Description of Subject}
  \item{Study.Group}{Description of Study.Group}
  \item{Age}{Description of Age}
  \item{Gender}{Description of Gender}
  \item{diet}{Description of diet}
  \item{Gender_binary}{Description of Gender_binary}
  \item{Diet_binary}{Description of Diet_binary}
  \item{Diet_bin1}{Description of Diet_bin1}
  \item{Diet_bin2}{Description of Diet_bin2}
}
}
\source{
He, Xuan, et al. "Fecal microbiome and metabolome of infants fed bovine MFGM supplemented formula or standard formula with breast-fed infants as reference: a randomized controlled trial." Scientific reports 9.1 (2019): 1-14
}
\usage{
He_2019_metadata
}
\description{
The data were pulled from this summarized article https://www.nature.com/articles/s41522-022-00345-5/figures/1, specifically reference the study by He, Xuan, et al. "Fecal microbiome and metabolome of infants fed bovine MFGM supplemented formula or standard formula with breast-fed infants as reference: a randomized controlled trial." Scientific reports 9.1 (2019): 1-14.
}
\keyword{datasets}
