% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIMON_Cyto_Export.R
\name{LIMON_Cyto_Export}
\alias{LIMON_Cyto_Export}
\title{Export Edge list for cytoscape}
\usage{
LIMON_Cyto_Export(Obj)
}
\arguments{
\item{Obj}{LIMON Object from running LIMON_Edges_Networks()}
}
\value{
Edge_Table written to Global Environment with Sources, Sinks and Edge Weights
}
\description{
Export Edge list for cytoscape
}
