% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIMON_Diff_Networks.R
\name{LIMON_Diff_Networks}
\alias{LIMON_Diff_Networks}
\title{Get graph network and edge table for difference between visits}
\usage{
LIMON_Diff_Networks(
  Obj,
  threshold = 0.2,
  vertex.size = 3,
  vertex.label.cex = 8,
  vertex.label.color = "black"
)
}
\arguments{
\item{Obj}{LIMON Object after performing LIMON_DistrFit()}

\item{threshold}{Numeric value of minimum value edges must be. Default is 0.2. Below that threshold will be converted to 0}

\item{vertex.size}{Numeric value for vertex size for network plotting. Default = 3}

\item{vertex.label.cex}{Numeric value for vertex label for network plotting. Default = 8}

\item{vertex.label.color}{Node label color. Default "black"}
}
\value{
A LIMON Object with DiffEdge_Table and Diff_Network added to it. DiffEdge_Table: List with DiffEdge_Table_i from CovMatrix_Diff_ji. Columns are sources and sinks (taxa rows and columns from covariance matrix) and Edge_weight - value from covariance matrix. Diff_Networks: A Network_Graph_ji igraph per difference in time points CovMatrix_Diff_ji. Prints a graph per time point
}
\description{
Get graph network and edge table for difference between visits
}
