% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIMON_DistrFit.R
\name{LIMON_DistrFit}
\alias{LIMON_DistrFit}
\title{Fit a zero inflated negative binomial linear mixed model with covariate correction}
\usage{
LIMON_DistrFit(
  Obj,
  Time,
  Subject,
  Covariates,
  model,
  prop.data = FALSE,
  correlation = NULL
)
}
\arguments{
\item{Obj}{(Required) A LIMON Object}

\item{Time}{(Required) A column in "SampleData" for the Time variable. Must be numeric ordinal (1,2,3, etc)}

\item{Subject}{(Required) A column of subject IDs. Used as}

\item{Covariates}{(Required) Specified string of columns from SampleData to use as covariates in the model}

\item{model}{(Required) Specify the right side of the lmer equation (the covariates to correct by)}

\item{prop.data}{(Optional) Option to specify if these are proprotional data or not. [Default] = FALSE indicating these are raw counts. If using relative abundance or other proportion data, specify TRUE.}

\item{correlation}{(Optional) Option to specify the correlation structure for model fitting ([Default] of NULL specifies unstructured). Follows the same covariance structure names as in NBZIMM}
}
\value{
LIMON Object with the addition; (1) Model_data: contains the Counts and SampleData in Long Format for the model. (2) equation: the formula for ZINBLMM. (3) Model_res: The results from running glmm.zinb(). (4) Corrected_Counts: corrected count values with samples as rows and taxa as columns. (5) Corrected_Counts_Time: corrected count values per time point with samples as rows and taxa as columns.
}
\description{
This function fits a zero inflated negative binomial linear mixed model with user specified Time variable, subjects as random effects, and covariates. The returned value are coutns per time point that have had the covariate effect removed from them. We recommend users to perform ZINBLMM model fitting and selection to determine the appropriate model and covariance structure before using this step. The function will fit a zero-inflated negative binomial linear mixed model with the covariates to correct by. The function will first convert the count table and sample data into a long formatted model table. The model will be fit with the outcome variable as "asvcounts" and the predictors as covariates to correct by with subjects modeled as random effects.
}
