% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIMON_Edges_Networks.R
\name{LIMON_Edges_Networks}
\alias{LIMON_Edges_Networks}
\title{Get edge table and printed network from LIMON after Network inference per time}
\usage{
LIMON_Edges_Networks(
  Obj,
  threshold = 0.2,
  vertex.size = 3,
  vertex.label.cex = 8,
  vertex.label.color = "black"
)
}
\arguments{
\item{Obj}{LIMON object from LIMON_Netnf_Time()}

\item{threshold}{Numeric value of minimum value edges must be. Default is 0.2. Below that threshold will be converted to 0}

\item{vertex.size}{Numeric value for vertex size for network plotting. Default = 3}

\item{vertex.label.cex}{Numeric value for vertex label for network plotting. Default = 8}

\item{vertex.label.color}{Node label color. Default "black"}
}
\value{
A LIMON Object with Edge_Table and Network_Graphs added to it. Edge_Table: List with Edge_Table_i for each Time point i. Columns are sources and sinks (taxa rows and columns from covariance matrix) and Edge_weight - value from covariance matrix. Network_Graphs: A Network_Graph_i igraph per time point i. Prints a graph per time point
}
\description{
Get edge table and printed network from LIMON after Network inference per time
}
