% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIMON_IndEdges.R
\name{LIMON_IndEdges}
\alias{LIMON_IndEdges}
\title{Extract edge table for Individual networks}
\usage{
LIMON_IndEdges(Obj, threshold = 0.2)
}
\arguments{
\item{Obj}{LIMON object from LIMON_IndNet()}

\item{threshold}{(Optional) absolute value of the minimum edge weight. [Default] = 0.2}
}
\value{
Returns a list of individual edge tables store in Individual_Edge_Table and a merged version with the Sample data in Merged_Edge_Table. This can be used for further analysis of edge changes overtime.
}
\description{
This function is the follow up to LIMON_IndNet(). After inferring individual networks (networks per sample), this function will extract the edge list for each unique pair of interactions.
}
