% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIMON_Obj.R
\name{LIMON_Obj}
\alias{LIMON_Obj}
\title{Counts to LIMON}
\usage{
LIMON_Obj(Counts, SampleData)
}
\arguments{
\item{Counts}{A genomic count table with samples as rows and taxa/genes/etc as columns}

\item{SampleData}{A sample data frame with row names matching "Counts"}
}
\value{
Returns a LIMON Object with the Counts and SampleData stored
}
\description{
Create Formatted Objects to run LIMON from raw count tables
}
