% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIMON_StatNodes.R
\name{LIMON_StatNodes}
\alias{LIMON_StatNodes}
\title{Individualized Microbial Node Variable selection}
\usage{
LIMON_StatNodes(
  LIMON_obj,
  dependent,
  node_type = "Edge_weight",
  pval = 1,
  adjpval = 1,
  time = "Time",
  method = "lm",
  timepoints = NULL,
  plot_nodes = FALSE,
  estimate = NULL,
  custom_colors = NULL,
  point_size = 7
)
}
\arguments{
\item{LIMON_obj}{A LIMON object after the LIMON_IndEdges() step}

\item{dependent}{User specified dependent variable. For Linear Regression, must be a continuous variable. For Logistic Regression, it must be binary (0,1). For Multinomial, it must have at least three categories.}

\item{node_type}{Option to specify what node characteristic should be used for variable selection. Default/only currently supported option is "Edge_weight"}

\item{pval}{Option to change the p value from the linear models to filter the models by. Default is no threshold applied (pvalue of 1).}

\item{adjpval}{Option to change the BH adjusted pvalue filtering threshold. Default is no threshold applied (adjusted pvalue of 1).}

\item{time}{Option to specify what the time column is labeled as. default is "Time"}

\item{method}{Default is "lm" for linear regression. Other options are "logit" for logistic regression or "multinom" for multinomial}

\item{timepoints}{Option to provide a vector of which timepoints to run the analysis on instead of all timepoints}

\item{plot_nodes}{Option to plot the significant interactions for linear or logistic models. Default FALSE}

\item{estimate}{Option of an absolute value to filter the interaction edge weight coefficient estimate by}

\item{custom_colors}{Option to specify either an R color palette or supply a vector of colors to use in the plot_nodes. Default is NULL and uses Set3 from RColorBrewer.}

\item{point_size}{Dot size for the plot option. Default 7}
}
\value{
A LIMON object with "Significant Interactions" which describes each model that passed thresholding, and "Nodes_Data" which will return a data table containing the metdata and only those taxa that were in the Significant_Interactions Table.
}
\description{
This function takes a LIMON output object and pulls out all the significant interactions by your dependent variable of interest. First the model will fit a linear regression with a user specified dependent variable, for every unique interaction between taxa. Then it will filter to the results only significant based on the users p-value threshold. These results will be saved to "Significant_Interactions".
}
