% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEC_Metagenomic_data}
\alias{NEC_Metagenomic_data}
\title{NEC subset of relative Metagenomic counts from "Necrotizing enterocolitis is preceded by increased gut bacterial replication, Klebsiella, and fimbriae-encoding bacteria." Olm et al. Sci Adv 2019;5(12):eaax5727}
\format{
A data frame with 405 rows and 360 variables:
\describe{
  \item{\code{Metagenomic_sequencing_assay_ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Sample_ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Participant_repeated_measure_ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Participant_ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.sp..M334..EUPATH_0009257_Bacteria_334.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.sp..F0442..EUPATH_0009257_Bacteria_42.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.sp..SK643..EUPATH_0009257_Bacteria_43.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.bolteae_CAG_59..EUPATH_0009257_Bacteria_59.}}{double COLUMN_DESCRIPTION}
  \item{\code{Pantoea.sp..PSNIH6..EUPATH_0009257_Bacteria_6.}}{double COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.oris..EUPATH_0009257_Bacteria_Actinobacteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_oris.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Rothia.dentocariosa..EUPATH_0009257_Bacteria_Actinobacteria_Actinobacteria_Micrococcales_Micrococcaceae_Rothia_Rothia_dentocariosa.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Rothia.mucilaginosa..EUPATH_0009257_Bacteria_Actinobacteria_Actinobacteria_Micrococcales_Micrococcaceae_Rothia_Rothia_mucilaginosa.}}{double COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.sp..oral.taxon.180..EUPATH_0009257_Bacteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_sp_oral_taxon_180.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.sp..oral.taxon.181..EUPATH_0009257_Bacteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_sp_oral_taxon_181.}}{double COLUMN_DESCRIPTION}
  \item{\code{Alloscardovia.omnicolens..EUPATH_0009257_Bacteria_Actinobacteria_Bifidobacteriales_Bifidobacteriaceae_Alloscardovia_Alloscardovia_omnicolens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bifidobacterium.bifidum..EUPATH_0009257_Bacteria_Actinobacteria_Bifidobacteriales_Bifidobacteriaceae_Bifidobacterium_Bifidobacterium_bifidum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bifidobacterium.dentium..EUPATH_0009257_Bacteria_Actinobacteria_Bifidobacteriales_Bifidobacteriaceae_Bifidobacterium_Bifidobacterium_dentium.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bifidobacterium.longum..EUPATH_0009257_Bacteria_Actinobacteria_Bifidobacteriales_Bifidobacteriaceae_Bifidobacterium_Bifidobacterium_longum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Atopobium.deltae..EUPATH_0009257_Bacteria_Actinobacteria_Coriobacteriia_Coriobacteriales_Atopobiaceae_Atopobium_Atopobium_deltae.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Atopobium.rimae..EUPATH_0009257_Bacteria_Actinobacteria_Coriobacteriia_Coriobacteriales_Atopobiaceae_Atopobium_Atopobium_rimae.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Enorma.massiliensis..EUPATH_0009257_Bacteria_Actinobacteria_Coriobacteriia_Coriobacteriales_Coriobacteriaceae_Enorma_Enorma_massiliensis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Slackia.exigua..EUPATH_0009257_Bacteria_Actinobacteria_Coriobacteriia_Eggerthellales_Eggerthellaceae_Slackia_Slackia_exigua.}}{double COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.coyleae..EUPATH_0009257_Bacteria_Actinobacteria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_coyleae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.hadale..EUPATH_0009257_Bacteria_Actinobacteria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_hadale.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Mycobacteroides.chelonae..EUPATH_0009257_Bacteria_Actinobacteria_Corynebacteriales_Mycobacteriaceae_Mycobacteroides_Mycobacteroides_chelonae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Cutibacterium.avidum..EUPATH_0009257_Bacteria_Actinobacteria_Propionibacteriales_Propionibacteriaceae_Cutibacterium_Cutibacterium_avidum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.caccae..EUPATH_0009257_Bacteria_Bacteroidetes_Bacteroidia_Bacteroidales_Bacteroidaceae_Bacteroides_Bacteroides_caccae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.dorei..EUPATH_0009257_Bacteria_Bacteroidetes_Bacteroidia_Bacteroidales_Bacteroidaceae_Bacteroides_Bacteroides_dorei.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.faecis..EUPATH_0009257_Bacteria_Bacteroidetes_Bacteroidia_Bacteroidales_Bacteroidaceae_Bacteroides_Bacteroides_faecis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.massiliensis..EUPATH_0009257_Bacteria_Bacteroidetes_Bacteroidia_Bacteroidales_Bacteroidaceae_Bacteroides_Bacteroides_massiliensis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.nordii..EUPATH_0009257_Bacteria_Bacteroidetes_Bacteroidia_Bacteroidales_Bacteroidaceae_Bacteroides_Bacteroides_nordii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.ovatus..EUPATH_0009257_Bacteria_Bacteroidetes_Bacteroidia_Bacteroidales_Bacteroidaceae_Bacteroides_Bacteroides_ovatus.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Prevotella.bivia..EUPATH_0009257_Bacteria_Bacteroidetes_Bacteroidia_Bacteroidales_Prevotellaceae_Prevotella_Prevotella_bivia.}}{double COLUMN_DESCRIPTION}
  \item{\code{Prevotella.buccalis..EUPATH_0009257_Bacteria_Bacteroidetes_Bacteroidia_Bacteroidales_Prevotellaceae_Prevotella_Prevotella_buccalis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Prevotella.copri..EUPATH_0009257_Bacteria_Bacteroidetes_Bacteroidia_Bacteroidales_Prevotellaceae_Prevotella_Prevotella_copri.}}{double COLUMN_DESCRIPTION}
  \item{\code{Prevotella.melaninogenica..EUPATH_0009257_Bacteria_Bacteroidetes_Bacteroidia_Bacteroidales_Prevotellaceae_Prevotella_Prevotella_melaninogenica.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Alistipes.putredinis..EUPATH_0009257_Bacteria_Bacteroidetes_Bacteroidia_Bacteroidales_Rikenellaceae_Alistipes_Alistipes_putredinis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Lachnoanaerobaculum.saburreum..EUPATH_0009257_Bacteria_Clostridia_Clostridiales_Lachnospiraceae_Lachnoanaerobaculum_Lachnoanaerobaculum_saburreum.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Paeniclostridium.sordellii..EUPATH_0009257_Bacteria_Clostridia_Clostridiales_Peptostreptococcaceae_Paeniclostridium_Paeniclostridium_sordellii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Paraclostridium.bifermentans..EUPATH_0009257_Bacteria_Clostridia_Clostridiales_Peptostreptococcaceae_Paraclostridium_Paraclostridium_bifermentans.}}{double COLUMN_DESCRIPTION}
  \item{\code{Terrisporobacter.glycolicus..EUPATH_0009257_Bacteria_Clostridia_Clostridiales_Peptostreptococcaceae_Terrisporobacter_Terrisporobacter_glycolicus.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Mogibacterium.diversum..EUPATH_0009257_Bacteria_Clostridiales_Clostridiales_Family_XIII_Incertae_Sedis_Mogibacterium_Mogibacterium_diversum.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Mogibacterium.pumilum..EUPATH_0009257_Bacteria_Clostridiales_Clostridiales_Family_XIII_Incertae_Sedis_Mogibacterium_Mogibacterium_pumilum.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Mogibacterium.timidum..EUPATH_0009257_Bacteria_Clostridiales_Clostridiales_Family_XIII_Incertae_Sedis_Mogibacterium_Mogibacterium_timidum.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Clostridium.sp..7_2_43FAA..EUPATH_0009257_Bacteria_FAA.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacillus.cereus.group..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Bacillaceae_Bacillus_Bacillus_cereus_group.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Bacillus.haynesii..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Bacillaceae_Bacillus_Bacillus_haynesii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Bacillus.hisashii..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Bacillaceae_Bacillus_Bacillus_hisashii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Bacillus.intestinalis..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Bacillaceae_Bacillus_Bacillus_intestinalis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacillus.thermoamylovorans..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Bacillaceae_Bacillus_Bacillus_thermoamylovorans.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Lysinibacillus.contaminans..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Bacillaceae_Lysinibacillus_Lysinibacillus_contaminans.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Gemella.asaccharolytica..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Bacillales_unclassified_Gemella_Gemella_asaccharolytica.}}{double COLUMN_DESCRIPTION}
  \item{\code{Gemella.haemolysans..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Bacillales_unclassified_Gemella_Gemella_haemolysans.}}{double COLUMN_DESCRIPTION}
  \item{\code{Gemella.morbillorum..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Bacillales_unclassified_Gemella_Gemella_morbillorum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Gemella.sanguinis..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Bacillales_unclassified_Gemella_Gemella_sanguinis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Paenibacillus.macerans..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Paenibacillaceae_Paenibacillus_Paenibacillus_macerans.}}{double COLUMN_DESCRIPTION}
  \item{\code{Staphylococcus.aureus..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Staphylococcaceae_Staphylococcus_Staphylococcus_aureus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Staphylococcus.capitis..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Staphylococcaceae_Staphylococcus_Staphylococcus_capitis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Staphylococcus.caprae..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Staphylococcaceae_Staphylococcus_Staphylococcus_caprae.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Staphylococcus.hominis..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Staphylococcaceae_Staphylococcus_Staphylococcus_hominis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Staphylococcus.simulans..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Staphylococcaceae_Staphylococcus_Staphylococcus_simulans.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Staphylococcus.warneri..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Bacillales_Staphylococcaceae_Staphylococcus_Staphylococcus_warneri.}}{double COLUMN_DESCRIPTION}
  \item{\code{Enterococcus.avium..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Enterococcaceae_Enterococcus_Enterococcus_avium.}}{double COLUMN_DESCRIPTION}
  \item{\code{Enterococcus.devriesei..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Enterococcaceae_Enterococcus_Enterococcus_devriesei.}}{double COLUMN_DESCRIPTION}
  \item{\code{Enterococcus.dispar..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Enterococcaceae_Enterococcus_Enterococcus_dispar.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Enterococcus.durans..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Enterococcaceae_Enterococcus_Enterococcus_durans.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Enterococcus.faecalis..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Enterococcaceae_Enterococcus_Enterococcus_faecalis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Enterococcus.faecium..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Enterococcaceae_Enterococcus_Enterococcus_faecium.}}{double COLUMN_DESCRIPTION}
  \item{\code{Enterococcus.gilvus..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Enterococcaceae_Enterococcus_Enterococcus_gilvus.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Enterococcus.hirae..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Enterococcaceae_Enterococcus_Enterococcus_hirae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Lactobacillus.oris..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Lactobacillaceae_Lactobacillus_Lactobacillus_oris.}}{double COLUMN_DESCRIPTION}
  \item{\code{Lactobacillus.paragasseri..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Lactobacillaceae_Lactobacillus_Lactobacillus_paragasseri.}}{double COLUMN_DESCRIPTION}
  \item{\code{Weissella.confusa..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Leuconostocaceae_Weissella_Weissella_confusa.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Lactococcus.lactis..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Streptococcaceae_Lactococcus_Lactococcus_lactis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.infantarius..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_infantarius.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.lutetiensis..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_lutetiensis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.mitis..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_mitis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.mutans..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_mutans.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.oralis..EUPATH_0009257_Bacteria_Firmicutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_oralis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.baratii..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_baratii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.botulinum..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_botulinum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.butyricum..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_butyricum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.cadaveris..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_cadaveris.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.celatum..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_celatum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.cochlearium..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_cochlearium.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.disporicum..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_disporicum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.neonatale..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_neonatale.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.perfringens..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_perfringens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.sp..CAG_58..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_sp_CAG_58.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Clostridium.sp..MSTE9..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_sp_MSTE9.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Clostridium.sporogenes..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_sporogenes.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.tepidum..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Clostridium_Clostridium_tepidum.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Hungatella.hathewayi..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Clostridiaceae_Hungatella_Hungatella_hathewayi.}}{double COLUMN_DESCRIPTION}
  \item{\code{Eubacterium.eligens..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Eubacteriaceae_Eubacterium_Eubacterium_eligens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Eubacterium.sp..CAG_38..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Eubacteriaceae_Eubacterium_Eubacterium_sp_CAG_38.}}{double COLUMN_DESCRIPTION}
  \item{\code{Blautia.coccoides..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Lachnospiraceae_Blautia_Blautia_coccoides.}}{double COLUMN_DESCRIPTION}
  \item{\code{Blautia.sp..CAG_257..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Lachnospiraceae_Blautia_Blautia_sp_CAG_257.}}{double COLUMN_DESCRIPTION}
  \item{\code{Blautia.wexlerae..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Lachnospiraceae_Blautia_Blautia_wexlerae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Ruminococcus.gnavus..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Lachnospiraceae_Blautia_Ruminococcus_gnavus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Ruminococcus.torques..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Lachnospiraceae_Blautia_Ruminococcus_torques.}}{double COLUMN_DESCRIPTION}
  \item{\code{Coprococcus.comes..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Lachnospiraceae_Coprococcus_Coprococcus_comes.}}{double COLUMN_DESCRIPTION}
  \item{\code{Dorea.longicatena..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Lachnospiraceae_Dorea_Dorea_longicatena.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.symbiosum..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Lachnospiraceae_Lachnoclostridium_Clostridium_symbiosum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Roseburia.faecis..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Lachnospiraceae_Roseburia_Roseburia_faecis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Roseburia.hominis..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Lachnospiraceae_Roseburia_Roseburia_hominis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Tyzzerella.nexilis..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Lachnospiraceae_Tyzzerella_Tyzzerella_nexilis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Peptococcus.niger..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Peptococcaceae_Peptococcus_Peptococcus_niger.}}{double COLUMN_DESCRIPTION}
  \item{\code{Anaerotruncus.sp..CAG_528..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Ruminococcaceae_Anaerotruncus_Anaerotruncus_sp_CAG_528.}}{double COLUMN_DESCRIPTION}
  \item{\code{Ruminococcus.bromii..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Ruminococcaceae_Ruminococcus_Ruminococcus_bromii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Ruminococcus.lactaris..EUPATH_0009257_Bacteria_Firmicutes_Clostridia_Clostridiales_Ruminococcaceae_Ruminococcus_Ruminococcus_lactaris.}}{double COLUMN_DESCRIPTION}
  \item{\code{Veillonella.dispar..EUPATH_0009257_Bacteria_Firmicutes_Negativicutes_Veillonellales_Veillonellaceae_Veillonella_Veillonella_dispar.}}{double COLUMN_DESCRIPTION}
  \item{\code{Anaerococcus.vaginalis..EUPATH_0009257_Bacteria_Firmicutes_Tissierellia_Tissierellales_Peptoniphilaceae_Anaerococcus_Anaerococcus_vaginalis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Finegoldia.magna..EUPATH_0009257_Bacteria_Firmicutes_Tissierellia_Tissierellales_Peptoniphilaceae_Finegoldia_Finegoldia_magna.}}{double COLUMN_DESCRIPTION}
  \item{\code{Sneathia.amnii..EUPATH_0009257_Bacteria_Fusobacteria_Fusobacteriia_Fusobacteriales_Leptotrichiaceae_Sneathia_Sneathia_amnii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Klebsiella.michiganensis..EUPATH_0009257_Bacteria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Klebsiella_Klebsiella_michiganensis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Leclercia.adecarboxylata..EUPATH_0009257_Bacteria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Leclercia_Leclercia_adecarboxylata.}}{double COLUMN_DESCRIPTION}
  \item{\code{Haemophilus.parahaemolyticus..EUPATH_0009257_Bacteria_Gammaproteobacteria_Pasteurellales_Pasteurellaceae_Haemophilus_Haemophilus_parahaemolyticus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Acinetobacter.nosocomialis..EUPATH_0009257_Bacteria_Gammaproteobacteria_Pseudomonadales_Moraxellaceae_Acinetobacter_Acinetobacter_nosocomialis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Negativicoccus.succinicivorans..EUPATH_0009257_Bacteria_Negativicutes_Veillonellales_Veillonellaceae_Negativicoccus_Negativicoccus_succinicivorans.}}{double COLUMN_DESCRIPTION}
  \item{\code{Neisseria.flavescens..EUPATH_0009257_Bacteria_Proteobacteria_Betaproteobacteria_Neisseriales_Neisseriaceae_Neisseria_Neisseria_flavescens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Neisseria.sicca..EUPATH_0009257_Bacteria_Proteobacteria_Betaproteobacteria_Neisseriales_Neisseriaceae_Neisseria_Neisseria_sicca.}}{double COLUMN_DESCRIPTION}
  \item{\code{Mixta.calida..EUPATH_0009257_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Erwiniaceae_Mixta_Mixta_calida.}}{double COLUMN_DESCRIPTION}
  \item{\code{Pantoea.allii..EUPATH_0009257_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Erwiniaceae_Pantoea_Pantoea_allii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Pantoea.vagans..EUPATH_0009257_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Erwiniaceae_Pantoea_Pantoea_vagans.}}{double COLUMN_DESCRIPTION}
  \item{\code{Hafnia.alvei..EUPATH_0009257_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Hafniaceae_Hafnia_Hafnia_alvei.}}{double COLUMN_DESCRIPTION}
  \item{\code{Hafnia.paralvei..EUPATH_0009257_Bacteria_Proteobacteria_Gammaproteobacteria_Enterobacterales_Hafniaceae_Hafnia_Hafnia_paralvei.}}{double COLUMN_DESCRIPTION}
  \item{\code{Robinsoniella.sp..RHS..EUPATH_0009257_Bacteria_RHS.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Mycoplasma.hominis..EUPATH_0009257_Bacteria_Tenericutes_Mollicutes_Mycoplasmatales_Mycoplasmataceae_Mycoplasma_Mycoplasma_hominis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Ureaplasma.parvum..EUPATH_0009257_Bacteria_Tenericutes_Mollicutes_Mycoplasmatales_Mycoplasmataceae_Ureaplasma_Ureaplasma_parvum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Eggerthella.lenta..EUPATH_0009257_Bacteria_a.}}{double COLUMN_DESCRIPTION}
  \item{\code{Aeriscardovia.aeriphila..EUPATH_0009257_Bacteria_a_Actinobacteria_Bifidobacteriales_Bifidobacteriaceae_Aeriscardovia_Aeriscardovia_aeriphila.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bifidobacterium.breve..EUPATH_0009257_Bacteria_a_Actinobacteria_Bifidobacteriales_Bifidobacteriaceae_Bifidobacterium_Bifidobacterium_breve.}}{double COLUMN_DESCRIPTION}
  \item{\code{Cutibacterium.acnes..EUPATH_0009257_Bacteria_a_Actinobacteria_Propionibacteriales_Propionibacteriaceae_Cutibacterium_Cutibacterium_acnes.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bradyrhizobium.sp..BTAi1..EUPATH_0009257_Bacteria_a_Alphaproteobacteria_Rhizobiales_Bradyrhizobiaceae_Bradyrhizobium_Bradyrhizobium_sp_BTAi1.}}{double COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.pseudodiphtheriticum..EUPATH_0009257_Bacteria_a_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_pseudodiphtheriticum.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Citrobacter.pasteurii..EUPATH_0009257_Bacteria_a_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Citrobacter_Citrobacter_pasteurii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Citrobacter.werkmanii..EUPATH_0009257_Bacteria_a_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Citrobacter_Citrobacter_werkmanii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Cronobacter.sakazakii..EUPATH_0009257_Bacteria_a_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Cronobacter_Cronobacter_sakazakii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Haemophilus.parainfluenzae..EUPATH_0009257_Bacteria_a_Gammaproteobacteria_Pasteurellales_Pasteurellaceae_Haemophilus_Haemophilus_parainfluenzae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Pseudomonas.putida.group..EUPATH_0009257_Bacteria_a_Gammaproteobacteria_Pseudomonadales_Pseudomonadaceae_Pseudomonas_Pseudomonas_putida_group.}}{double COLUMN_DESCRIPTION}
  \item{\code{Propionimicrobium.lymphophilum..EUPATH_0009257_Bacteria_a_Propionibacteriales_Propionibacteriaceae_Propionimicrobium_Propionimicrobium_lymphophilum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.sp..HPA0247..EUPATH_0009257_Bacteria_acteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_sp_HPA0247.}}{double COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.sp..S6_Spd3..EUPATH_0009257_Bacteria_acteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_sp_S6_Spd3.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bifidobacterium.kashiwanohense..EUPATH_0009257_Bacteria_acteria_Bifidobacteriales_Bifidobacteriaceae_Bifidobacterium_Bifidobacterium_kashiwanohense.}}{double COLUMN_DESCRIPTION}
  \item{\code{Collinsella.stercoris..EUPATH_0009257_Bacteria_acteria_Coriobacteriia_Coriobacteriales_Coriobacteriaceae_Collinsella_Collinsella_stercoris.}}{double COLUMN_DESCRIPTION}
  \item{\code{X.Collinsella..massiliensis..EUPATH_0009257_Bacteria_acteria_Coriobacteriia_Coriobacteriales_Coriobacteriaceae_Enorma_Collinsella_massiliensis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Gordonibacter.pamelaeae..EUPATH_0009257_Bacteria_acteria_Coriobacteriia_Eggerthellales_Eggerthellaceae_Gordonibacter_Gordonibacter_pamelaeae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.kroppenstedtii..EUPATH_0009257_Bacteria_acteria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_kroppenstedtii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Shewanella.oneidensis..EUPATH_0009257_Bacteria_acteria_Gammaproteobacteria_Alteromonadales_Shewanellaceae_Shewanella_Shewanella_oneidensis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Kluyvera.ascorbata..EUPATH_0009257_Bacteria_acteria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Kluyvera_Kluyvera_ascorbata.}}{double COLUMN_DESCRIPTION}
  \item{\code{Kosakonia.cowanii..EUPATH_0009257_Bacteria_acteria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Kosakonia_Kosakonia_cowanii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Kosakonia.sp..S29..EUPATH_0009257_Bacteria_acteria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Kosakonia_Kosakonia_sp_S29.}}{double COLUMN_DESCRIPTION}
  \item{\code{Haemophilus.quentini..EUPATH_0009257_Bacteria_acteria_Gammaproteobacteria_Pasteurellales_Pasteurellaceae_Haemophilus_Haemophilus_quentini.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Acinetobacter.junii..EUPATH_0009257_Bacteria_acteria_Gammaproteobacteria_Pseudomonadales_Moraxellaceae_Acinetobacter_Acinetobacter_junii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Propionibacterium.namnetense..EUPATH_0009257_Bacteria_acteria_Propionibacteriales_Propionibacteriaceae_Cutibacterium_Propionibacterium_namnetense.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.xylanisolvens..EUPATH_0009257_Bacteria_acteroidetes_Bacteroidia_Bacteroidales_Bacteroidaceae_Bacteroides_Bacteroides_xylanisolvens.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Clostridium.bolteae..EUPATH_0009257_Bacteria_ae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Romboutsia.ilealis..EUPATH_0009257_Bacteria_alis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Citrobacter.amalonaticus..EUPATH_0009257_Bacteria_ammaproteobacteria_Enterobacterales_Enterobacteriaceae_Citrobacter_Citrobacter_amalonaticus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Acinetobacter.radioresistens..EUPATH_0009257_Bacteria_ammaproteobacteria_Pseudomonadales_Moraxellaceae_Acinetobacter_Acinetobacter_radioresistens.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Klebsiella.variicola_CAG_634..EUPATH_0009257_Bacteria_aproteobacteria_Enterobacterales_Enterobacteriaceae_Klebsiella_Klebsiella_variicola_CAG_634.}}{double COLUMN_DESCRIPTION}
  \item{\code{Haemophilus.paraphrohaemolyticus..EUPATH_0009257_Bacteria_aproteobacteria_Pasteurellales_Pasteurellaceae_Haemophilus_Haemophilus_paraphrohaemolyticus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Enterococcus.gallinarum..EUPATH_0009257_Bacteria_arum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.vulgatus..EUPATH_0009257_Bacteria_atus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.naeslundii..EUPATH_0009257_Bacteria_bacteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_naeslundii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.turicensis..EUPATH_0009257_Bacteria_bacteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_turicensis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Trueperella.bernardiae..EUPATH_0009257_Bacteria_bacteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Trueperella_Trueperella_bernardiae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Varibaculum.cambriense..EUPATH_0009257_Bacteria_bacteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Varibaculum_Varibaculum_cambriense.}}{double COLUMN_DESCRIPTION}
  \item{\code{Roseomonas.mucosa..EUPATH_0009257_Bacteria_bacteria_Alphaproteobacteria_Rhodospirillales_Acetobacteraceae_Roseomonas_Roseomonas_mucosa.}}{double COLUMN_DESCRIPTION}
  \item{\code{Slackia.isoflavoniconvertens..EUPATH_0009257_Bacteria_bacteria_Coriobacteriia_Eggerthellales_Eggerthellaceae_Slackia_Slackia_isoflavoniconvertens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Morganella.morganii..EUPATH_0009257_Bacteria_bacteria_Gammaproteobacteria_Enterobacterales_Morganellaceae_Morganella_Morganella_morganii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Gordonia.jacobaea..EUPATH_0009257_Bacteria_baea.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Akkermansia.muciniphila..EUPATH_0009257_Bacteria_bia_Verrucomicrobiae_Verrucomicrobiales_Akkermansiaceae_Akkermansia_Akkermansia_muciniphila.}}{double COLUMN_DESCRIPTION}
  \item{\code{Neisseria.macacae..EUPATH_0009257_Bacteria_cae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Neisseria.mucosa..EUPATH_0009257_Bacteria_cosa.}}{double COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.graevenitzii..EUPATH_0009257_Bacteria_cteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_graevenitzii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.urogenitalis..EUPATH_0009257_Bacteria_cteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_urogenitalis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Agrobacterium.tumefaciens..EUPATH_0009257_Bacteria_cteria_Alphaproteobacteria_Rhizobiales_Rhizobiaceae_Agrobacterium_Agrobacterium_tumefaciens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Enterorhabdus.caecimuris..EUPATH_0009257_Bacteria_cteria_Coriobacteriia_Eggerthellales_Eggerthellaceae_Enterorhabdus_Enterorhabdus_caecimuris.}}{double COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.kefirresidentii..EUPATH_0009257_Bacteria_cteria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_kefirresidentii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Shewanella.xiamenensis..EUPATH_0009257_Bacteria_cteria_Gammaproteobacteria_Alteromonadales_Shewanellaceae_Shewanella_Shewanella_xiamenensis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Escherichia.coli..EUPATH_0009257_Bacteria_cteria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Escherichia_Escherichia_coli.}}{double COLUMN_DESCRIPTION}
  \item{\code{Kluyvera.intermedia..EUPATH_0009257_Bacteria_cteria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Kluyvera_Kluyvera_intermedia.}}{double COLUMN_DESCRIPTION}
  \item{\code{Acinetobacter.pittii..EUPATH_0009257_Bacteria_cteria_Gammaproteobacteria_Pseudomonadales_Moraxellaceae_Acinetobacter_Acinetobacter_pittii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Pseudomonas.helleri..EUPATH_0009257_Bacteria_cteria_Gammaproteobacteria_Pseudomonadales_Pseudomonadaceae_Pseudomonas_Pseudomonas_helleri.}}{double COLUMN_DESCRIPTION}
  \item{\code{Parabacteroides.merdae..EUPATH_0009257_Bacteria_cteroidetes_Bacteroidia_Bacteroidales_Tannerellaceae_Parabacteroides_Parabacteroides_merdae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.neuii..EUPATH_0009257_Bacteria_ctinobacteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_neuii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Dermabacter.hominis..EUPATH_0009257_Bacteria_ctinobacteria_Actinobacteria_Micrococcales_Dermabacteraceae_Dermabacter_Dermabacter_hominis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Micrococcus.aloeverae..EUPATH_0009257_Bacteria_ctinobacteria_Actinobacteria_Micrococcales_Micrococcaceae_Micrococcus_Micrococcus_aloeverae.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Bifidobacterium.animalis..EUPATH_0009257_Bacteria_ctinobacteria_Bifidobacteriales_Bifidobacteriaceae_Bifidobacterium_Bifidobacterium_animalis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.accolens..EUPATH_0009257_Bacteria_ctinobacteria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_accolens.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.simulans..EUPATH_0009257_Bacteria_ctinobacteria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_simulans.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.pseudopneumoniae..EUPATH_0009257_Bacteria_cutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_pseudopneumoniae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Peptoniphilus.sp..DNF00840..EUPATH_0009257_Bacteria_cutes_Tissierellia_Tissierellales_Peptoniphilaceae_Peptoniphilus_Peptoniphilus_sp_DNF00840.}}{double COLUMN_DESCRIPTION}
  \item{\code{Aeromonas.caviae..EUPATH_0009257_Bacteria_e.}}{double COLUMN_DESCRIPTION}
  \item{\code{Peptoniphilus.harei..EUPATH_0009257_Bacteria_ei.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.plebeius..EUPATH_0009257_Bacteria_eius.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.pyogenes..EUPATH_0009257_Bacteria_enes.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.argentinense..EUPATH_0009257_Bacteria_ense.}}{double COLUMN_DESCRIPTION}
  \item{\code{Serratia.liquefaciens..EUPATH_0009257_Bacteria_eobacteria_Gammaproteobacteria_Enterobacterales_Yersiniaceae_Serratia_Serratia_liquefaciens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Pantoea.brenneri..EUPATH_0009257_Bacteria_eri.}}{double COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.sp..HMSC035G02..EUPATH_0009257_Bacteria_eria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_sp_HMSC035G02.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bifidobacterium.pseudocatenulatum..EUPATH_0009257_Bacteria_eria_Bifidobacteriales_Bifidobacteriaceae_Bifidobacterium_Bifidobacterium_pseudocatenulatum.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.pyruviciproducens..EUPATH_0009257_Bacteria_eria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_pyruviciproducens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Citrobacter.koseri..EUPATH_0009257_Bacteria_eria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Citrobacter_Citrobacter_koseri.}}{double COLUMN_DESCRIPTION}
  \item{\code{Enterobacter.mori..EUPATH_0009257_Bacteria_eria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Enterobacter_Enterobacter_mori.}}{double COLUMN_DESCRIPTION}
  \item{\code{Kluyvera.cryocrescens..EUPATH_0009257_Bacteria_eria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Kluyvera_Kluyvera_cryocrescens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Acinetobacter.ursingii..EUPATH_0009257_Bacteria_eria_Gammaproteobacteria_Pseudomonadales_Moraxellaceae_Acinetobacter_Acinetobacter_ursingii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.cellulosilyticus..EUPATH_0009257_Bacteria_eroidetes_Bacteroidia_Bacteroidales_Bacteroidaceae_Bacteroides_Bacteroides_cellulosilyticus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.thetaiotaomicron..EUPATH_0009257_Bacteria_eroidetes_Bacteroidia_Bacteroidales_Bacteroidaceae_Bacteroides_Bacteroides_thetaiotaomicron.}}{double COLUMN_DESCRIPTION}
  \item{\code{Prevotella.sp..oral.taxon.306..EUPATH_0009257_Bacteria_eroidetes_Bacteroidia_Bacteroidales_Prevotellaceae_Prevotella_Prevotella_sp_oral_taxon_306.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Capnocytophaga.ochracea..EUPATH_0009257_Bacteria_es_Flavobacteriia_Flavobacteriales_Flavobacteriaceae_Capnocytophaga_Capnocytophaga_ochracea.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.eggerthii..EUPATH_0009257_Bacteria_hii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Staphylococcus.pettenkoferi..EUPATH_0009257_Bacteria_i.}}{double COLUMN_DESCRIPTION}
  \item{\code{Aeromonas.media..EUPATH_0009257_Bacteria_ia.}}{double COLUMN_DESCRIPTION}
  \item{\code{Fusobacterium.periodonticum..EUPATH_0009257_Bacteria_ia_Fusobacteriia_Fusobacteriales_Fusobacteriaceae_Fusobacterium_Fusobacterium_periodonticum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Citrobacter.freundii..EUPATH_0009257_Bacteria_ia_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Citrobacter_Citrobacter_freundii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Klebsiella.pneumoniae..EUPATH_0009257_Bacteria_ia_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Klebsiella_Klebsiella_pneumoniae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Kluyvera.intestini..EUPATH_0009257_Bacteria_ia_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Metakosakonia_Kluyvera_intestini.}}{double COLUMN_DESCRIPTION}
  \item{\code{Raoultella.planticola..EUPATH_0009257_Bacteria_ia_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Raoultella_Raoultella_planticola.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Haemophilus.sp..HMSC71H05..EUPATH_0009257_Bacteria_ia_Gammaproteobacteria_Pasteurellales_Pasteurellaceae_Haemophilus_Haemophilus_sp_HMSC71H05.}}{double COLUMN_DESCRIPTION}
  \item{\code{Acinetobacter.bereziniae..EUPATH_0009257_Bacteria_ia_Gammaproteobacteria_Pseudomonadales_Moraxellaceae_Acinetobacter_Acinetobacter_bereziniae.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Acinetobacter.guillouiae..EUPATH_0009257_Bacteria_ia_Gammaproteobacteria_Pseudomonadales_Moraxellaceae_Acinetobacter_Acinetobacter_guillouiae.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Clostridium.amylolyticum..EUPATH_0009257_Bacteria_icum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Leuconostoc.pseudomesenteroides..EUPATH_0009257_Bacteria_icutes_Bacilli_Lactobacillales_Leuconostocaceae_Leuconostoc_Leuconostoc_pseudomesenteroides.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.anginosus.group..EUPATH_0009257_Bacteria_icutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_anginosus_group.}}{double COLUMN_DESCRIPTION}
  \item{\code{Megasphaera.micronuciformis..EUPATH_0009257_Bacteria_icutes_Negativicutes_Veillonellales_Veillonellaceae_Megasphaera_Megasphaera_micronuciformis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Lactobacillus.johnsonii..EUPATH_0009257_Bacteria_ii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.fragilis..EUPATH_0009257_Bacteria_ilis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Romboutsia.hominis..EUPATH_0009257_Bacteria_inis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Dermabacter.vaginalis..EUPATH_0009257_Bacteria_inobacteria_Actinobacteria_Micrococcales_Dermabacteraceae_Dermabacter_Dermabacter_vaginalis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.amycolatum..EUPATH_0009257_Bacteria_inobacteria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_amycolatum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Cutibacterium.granulosum..EUPATH_0009257_Bacteria_inobacteria_Propionibacteriales_Propionibacteriaceae_Cutibacterium_Cutibacterium_granulosum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.pasteurianus..EUPATH_0009257_Bacteria_irmicutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_pasteurianus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.thermophilus..EUPATH_0009257_Bacteria_irmicutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_thermophilus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.vestibularis..EUPATH_0009257_Bacteria_irmicutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_vestibularis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Megasphaera.sp..MJR8396C..EUPATH_0009257_Bacteria_irmicutes_Negativicutes_Veillonellales_Veillonellaceae_Megasphaera_Megasphaera_sp_MJR8396C.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Veillonella.sp..T11011_6..EUPATH_0009257_Bacteria_irmicutes_Negativicutes_Veillonellales_Veillonellaceae_Veillonella_Veillonella_sp_T11011_6.}}{double COLUMN_DESCRIPTION}
  \item{\code{Veillonella.tobetsuensis..EUPATH_0009257_Bacteria_irmicutes_Negativicutes_Veillonellales_Veillonellaceae_Veillonella_Veillonella_tobetsuensis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Staphylococcus.epidermidis..EUPATH_0009257_Bacteria_is.}}{double COLUMN_DESCRIPTION}
  \item{\code{Cloacibacterium.normanense..EUPATH_0009257_Bacteria_lavobacteriia_Flavobacteriales_Flavobacteriaceae_Cloacibacterium_Cloacibacterium_normanense.}}{double COLUMN_DESCRIPTION}
  \item{\code{Lactobacillus.animalis..EUPATH_0009257_Bacteria_lis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Veillonellaceae.bacterium_DNF00626..EUPATH_0009257_Bacteria_llonellales_Veillonellaceae_Veillonellaceae_unclassified_Veillonellaceae_bacterium_DNF00626.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Campylobacter.concisus..EUPATH_0009257_Bacteria_lonproteobacteria_Campylobacterales_Campylobacteraceae_Campylobacter_Campylobacter_concisus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Paraclostridium.benzoelyticum..EUPATH_0009257_Bacteria_lostridia_Clostridiales_Peptostreptococcaceae_Paraclostridium_Paraclostridium_benzoelyticum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Terrisporobacter.othiniensis..EUPATH_0009257_Bacteria_lostridia_Clostridiales_Peptostreptococcaceae_Terrisporobacter_Terrisporobacter_othiniensis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Brevundimonas.vesicularis..EUPATH_0009257_Bacteria_lphaproteobacteria_Caulobacterales_Caulobacteraceae_Brevundimonas_Brevundimonas_vesicularis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bradyrhizobium.sp..STM_3809..EUPATH_0009257_Bacteria_lphaproteobacteria_Rhizobiales_Bradyrhizobiaceae_Bradyrhizobium_Bradyrhizobium_sp_STM_3809.}}{double COLUMN_DESCRIPTION}
  \item{\code{Atopobium.parvulum..EUPATH_0009257_Bacteria_lum.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Franconibacter.pulveris..EUPATH_0009257_Bacteria_maproteobacteria_Enterobacterales_Enterobacteriaceae_Franconibacter_Franconibacter_pulveris.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Aggregatibacter.aphrophilus..EUPATH_0009257_Bacteria_maproteobacteria_Pasteurellales_Pasteurellaceae_Aggregatibacter_Aggregatibacter_aphrophilus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Pseudomonas.fluorescens.group..EUPATH_0009257_Bacteria_maproteobacteria_Pseudomonadales_Pseudomonadaceae_Pseudomonas_Pseudomonas_fluorescens_group.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.sp..HMSC034E03..EUPATH_0009257_Bacteria_micutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_sp_HMSC034E03.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.sp..HMSC070B10..EUPATH_0009257_Bacteria_micutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_sp_HMSC070B10.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.sp..HMSC071D03..EUPATH_0009257_Bacteria_micutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_sp_HMSC071D03.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Eisenbergiella.massiliensis..EUPATH_0009257_Bacteria_micutes_Clostridia_Clostridiales_Lachnospiraceae_Eisenbergiella_Eisenbergiella_massiliensis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Anaerococcus.hydrogenalis..EUPATH_0009257_Bacteria_micutes_Tissierellia_Tissierellales_Peptoniphilaceae_Anaerococcus_Anaerococcus_hydrogenalis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Peptoniphilus.lacrimalis..EUPATH_0009257_Bacteria_micutes_Tissierellia_Tissierellales_Peptoniphilaceae_Peptoniphilus_Peptoniphilus_lacrimalis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.uniformis..EUPATH_0009257_Bacteria_mis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Atlantibacter.hermannii..EUPATH_0009257_Bacteria_mmaproteobacteria_Enterobacterales_Enterobacteriaceae_Atlantibacter_Atlantibacter_hermannii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Citrobacter.portucalensis..EUPATH_0009257_Bacteria_mmaproteobacteria_Enterobacterales_Enterobacteriaceae_Citrobacter_Citrobacter_portucalensis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Klebsiella.quasipneumoniae..EUPATH_0009257_Bacteria_mmaproteobacteria_Enterobacterales_Enterobacteriaceae_Klebsiella_Klebsiella_quasipneumoniae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Pluralibacter.gergoviae..EUPATH_0009257_Bacteria_mmaproteobacteria_Enterobacterales_Enterobacteriaceae_Pluralibacter_Pluralibacter_gergoviae.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Raoultella.ornithinolytica..EUPATH_0009257_Bacteria_mmaproteobacteria_Enterobacterales_Enterobacteriaceae_Raoultella_Raoultella_ornithinolytica.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Pseudomonas.aeruginosa.group..EUPATH_0009257_Bacteria_mmaproteobacteria_Pseudomonadales_Pseudomonadaceae_Pseudomonas_Pseudomonas_aeruginosa_group.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.gordonii..EUPATH_0009257_Bacteria_nii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.radingae..EUPATH_0009257_Bacteria_nobacteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_radingae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.afermentans..EUPATH_0009257_Bacteria_nobacteria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_afermentans.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Prevotella.timonensis..EUPATH_0009257_Bacteria_nsis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.europaeus..EUPATH_0009257_Bacteria_obacteria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_europaeus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Lautropia.mirabilis..EUPATH_0009257_Bacteria_obacteria_Betaproteobacteria_Burkholderiales_Burkholderiaceae_Lautropia_Lautropia_mirabilis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Bifidobacterium.adolescentis..EUPATH_0009257_Bacteria_obacteria_Bifidobacteriales_Bifidobacteriaceae_Bifidobacterium_Bifidobacterium_adolescentis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Fusobacterium.varium..EUPATH_0009257_Bacteria_obacteria_Fusobacteriia_Fusobacteriales_Fusobacteriaceae_Fusobacterium_Fusobacterium_varium.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Oscillatoria.acuminata..EUPATH_0009257_Bacteria_obacteria_unclassified_Oscillatoriales_Oscillatoriaceae_Oscillatoria_Oscillatoria_acuminata.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Parabacteroides.distasonis..EUPATH_0009257_Bacteria_oidetes_Bacteroidia_Bacteroidales_Tannerellaceae_Parabacteroides_Parabacteroides_distasonis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.coprocola..EUPATH_0009257_Bacteria_ola.}}{double COLUMN_DESCRIPTION}
  \item{\code{Denitrobacterium.detoxificans..EUPATH_0009257_Bacteria_oriobacteriia_Eggerthellales_Eggerthellaceae_Denitrobacterium_Denitrobacterium_detoxificans.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.peroris..EUPATH_0009257_Bacteria_oris.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Veillonella.rogosae..EUPATH_0009257_Bacteria_osae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Enterococcus.raffinosus..EUPATH_0009257_Bacteria_osus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Pseudescherichia.vulneris..EUPATH_0009257_Bacteria_oteobacteria_Enterobacterales_Enterobacteriaceae_Pseudescherichia_Pseudescherichia_vulneris.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Aeromonas.dhakensis..EUPATH_0009257_Bacteria_oteobacteria_Gammaproteobacteria_Aeromonadales_Aeromonadaceae_Aeromonas_Aeromonas_dhakensis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Serratia.marcescens..EUPATH_0009257_Bacteria_oteobacteria_Gammaproteobacteria_Enterobacterales_Yersiniaceae_Serratia_Serratia_marcescens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Stenotrophomonas.maltophilia..EUPATH_0009257_Bacteria_oteobacteria_Xanthomonadales_Xanthomonadaceae_Stenotrophomonas_Stenotrophomonas_maltophilia.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.innocuum..EUPATH_0009257_Bacteria_otrichia_Erysipelotrichales_Erysipelotrichaceae_Erysipelatoclostridium_Clostridium_innocuum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Sphingomonas.paucimobilis..EUPATH_0009257_Bacteria_phaproteobacteria_Sphingomonadales_Sphingomonadaceae_Sphingomonas_Sphingomonas_paucimobilis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Veillonella.atypica..EUPATH_0009257_Bacteria_pica.}}{double COLUMN_DESCRIPTION}
  \item{\code{Campylobacter.ureolyticus..EUPATH_0009257_Bacteria_proteobacteria_Campylobacterales_Campylobacteraceae_Campylobacter_Campylobacter_ureolyticus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Franconibacter.helveticus..EUPATH_0009257_Bacteria_proteobacteria_Enterobacterales_Enterobacteriaceae_Franconibacter_Franconibacter_helveticus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Neisseria.cinerea..EUPATH_0009257_Bacteria_rea.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Adlercreutzia.equolifaciens..EUPATH_0009257_Bacteria_ria_Coriobacteriia_Eggerthellales_Eggerthellaceae_Adlercreutzia_Adlercreutzia_equolifaciens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.tuberculostearicum..EUPATH_0009257_Bacteria_ria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_tuberculostearicum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Yersiniaceae.bacterium_2016Iso3..EUPATH_0009257_Bacteria_ria_Enterobacterales_Yersiniaceae_Yersiniaceae_unclassified_Yersiniaceae_bacterium_2016Iso3.}}{double COLUMN_DESCRIPTION}
  \item{\code{Yersiniaceae.bacterium_2016Iso4..EUPATH_0009257_Bacteria_ria_Enterobacterales_Yersiniaceae_Yersiniaceae_unclassified_Yersiniaceae_bacterium_2016Iso4.}}{double COLUMN_DESCRIPTION}
  \item{\code{Citrobacter.braakii..EUPATH_0009257_Bacteria_ria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Citrobacter_Citrobacter_braakii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Citrobacter.youngae..EUPATH_0009257_Bacteria_ria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Citrobacter_Citrobacter_youngae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Klebsiella.aerogenes..EUPATH_0009257_Bacteria_ria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Klebsiella_Klebsiella_aerogenes.}}{double COLUMN_DESCRIPTION}
  \item{\code{Klebsiella.grimontii..EUPATH_0009257_Bacteria_ria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Klebsiella_Klebsiella_grimontii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Klebsiella.variicola..EUPATH_0009257_Bacteria_ria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Klebsiella_Klebsiella_variicola.}}{double COLUMN_DESCRIPTION}
  \item{\code{Haemophilus.haemolyticus..EUPATH_0009257_Bacteria_ria_Gammaproteobacteria_Pasteurellales_Pasteurellaceae_Haemophilus_Haemophilus_haemolyticus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Acinetobacter.baumannii..EUPATH_0009257_Bacteria_ria_Gammaproteobacteria_Pseudomonadales_Moraxellaceae_Acinetobacter_Acinetobacter_baumannii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Acinetobacter.johnsonii..EUPATH_0009257_Bacteria_ria_Gammaproteobacteria_Pseudomonadales_Moraxellaceae_Acinetobacter_Acinetobacter_johnsonii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Bacteroides.stercoris..EUPATH_0009257_Bacteria_ris.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.parasanguinis..EUPATH_0009257_Bacteria_rmicutes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_parasanguinis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Anaerocolumna.aminovalerica..EUPATH_0009257_Bacteria_rmicutes_Clostridia_Clostridiales_Lachnospiraceae_Anaerocolumna_Anaerocolumna_aminovalerica.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Acidovorax.caeni..EUPATH_0009257_Bacteria_roteobacteria_Betaproteobacteria_Burkholderiales_Comamonadaceae_Acidovorax_Acidovorax_caeni.}}{double COLUMN_DESCRIPTION}
  \item{\code{Enterobacter.cloacae.complex..EUPATH_0009257_Bacteria_roteobacteria_Enterobacterales_Enterobacteriaceae_Enterobacter_Enterobacter_cloacae_complex.}}{double COLUMN_DESCRIPTION}
  \item{\code{Proteus.mirabilis..EUPATH_0009257_Bacteria_roteobacteria_Gammaproteobacteria_Enterobacterales_Morganellaceae_Proteus_Proteus_mirabilis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Serratia.fonticola..EUPATH_0009257_Bacteria_roteobacteria_Gammaproteobacteria_Enterobacterales_Yersiniaceae_Serratia_Serratia_fonticola.}}{double COLUMN_DESCRIPTION}
  \item{\code{Stenotrophomonas.rhizophila..EUPATH_0009257_Bacteria_roteobacteria_Xanthomonadales_Xanthomonadaceae_Stenotrophomonas_Stenotrophomonas_rhizophila.}}{double COLUMN_DESCRIPTION}
  \item{\code{Pantoea.dispersa..EUPATH_0009257_Bacteria_rsa.}}{double COLUMN_DESCRIPTION}
  \item{\code{Erysipelatoclostridium.ramosum..EUPATH_0009257_Bacteria_rysipelotrichales_Erysipelotrichaceae_Erysipelatoclostridium_Erysipelatoclostridium_ramosum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Staphylococcus.haemolyticus..EUPATH_0009257_Bacteria_s.}}{double COLUMN_DESCRIPTION}
  \item{\code{Fusicatenibacter.saccharivorans..EUPATH_0009257_Bacteria_s_Clostridia_Clostridiales_Lachnospiraceae_Fusicatenibacter_Fusicatenibacter_saccharivorans.}}{double COLUMN_DESCRIPTION}
  \item{\code{Eubacterium.rectale..EUPATH_0009257_Bacteria_s_Clostridia_Clostridiales_Lachnospiraceae_Lachnospiraceae_unclassified_Eubacterium_rectale.}}{double COLUMN_DESCRIPTION}
  \item{\code{Intestinibacter.bartlettii..EUPATH_0009257_Bacteria_s_Clostridia_Clostridiales_Peptostreptococcaceae_Intestinibacter_Intestinibacter_bartlettii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Ruthenibacterium.lactatiformans..EUPATH_0009257_Bacteria_s_Clostridia_Clostridiales_Ruminococcaceae_Ruthenibacterium_Ruthenibacterium_lactatiformans.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Lactobacillus.gasseri..EUPATH_0009257_Bacteria_seri.}}{double COLUMN_DESCRIPTION}
  \item{\code{Campylobacter.curvus..EUPATH_0009257_Bacteria_silonproteobacteria_Campylobacterales_Campylobacteraceae_Campylobacter_Campylobacter_curvus.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Lachnospiraceae.bacterium_2_1_46FAA..EUPATH_0009257_Bacteria_stridiales_Lachnospiraceae_Lachnospiraceae_unclassified_Lachnospiraceae_bacterium_2_1_46FAA.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Phytobacter.ursingii..EUPATH_0009257_Bacteria_teobacteria_Enterobacterales_Enterobacterales_unclassified_Phytobacter_Phytobacter_ursingii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Superficieibacter.electus..EUPATH_0009257_Bacteria_teobacteria_Enterobacterales_Enterobacteriaceae_Superficieibacter_Superficieibacter_electus.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Aeromonas.hydrophila..EUPATH_0009257_Bacteria_teobacteria_Gammaproteobacteria_Aeromonadales_Aeromonadaceae_Aeromonas_Aeromonas_hydrophila.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Moraxella.osloensis..EUPATH_0009257_Bacteria_teobacteria_Gammaproteobacteria_Pseudomonadales_Moraxellaceae_Moraxella_Moraxella_osloensis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Lactobacillus.reuteri..EUPATH_0009257_Bacteria_teri.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Actinomyces.odontolyticus..EUPATH_0009257_Bacteria_teria_Actinobacteria_Actinomycetales_Actinomycetaceae_Actinomyces_Actinomyces_odontolyticus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Collinsella.aerofaciens..EUPATH_0009257_Bacteria_teria_Coriobacteriia_Coriobacteriales_Coriobacteriaceae_Collinsella_Collinsella_aerofaciens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Asaccharobacter.celatus..EUPATH_0009257_Bacteria_teria_Coriobacteriia_Eggerthellales_Eggerthellaceae_Asaccharobacter_Asaccharobacter_celatus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.glucuronolyticum..EUPATH_0009257_Bacteria_teria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_glucuronolyticum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Corynebacterium.pseudogenitalium..EUPATH_0009257_Bacteria_teria_Corynebacteriales_Corynebacteriaceae_Corynebacterium_Corynebacterium_pseudogenitalium.}}{double COLUMN_DESCRIPTION}
  \item{\code{Lawsonella.clevelandensis..EUPATH_0009257_Bacteria_teria_Corynebacteriales_Corynebacteriales_unclassified_Lawsonella_Lawsonella_clevelandensis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Klebsiella.oxytoca..EUPATH_0009257_Bacteria_teria_Gammaproteobacteria_Enterobacterales_Enterobacteriaceae_Klebsiella_Klebsiella_oxytoca.}}{double COLUMN_DESCRIPTION}
  \item{\code{Haemophilus.influenzae..EUPATH_0009257_Bacteria_teria_Gammaproteobacteria_Pasteurellales_Pasteurellaceae_Haemophilus_Haemophilus_influenzae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Acinetobacter.lwoffii..EUPATH_0009257_Bacteria_teria_Gammaproteobacteria_Pseudomonadales_Moraxellaceae_Acinetobacter_Acinetobacter_lwoffii.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Enhydrobacter.aerosaccus..EUPATH_0009257_Bacteria_teria_Rhodospirillales_Rhodospirillales_unclassified_Enhydrobacter_Enhydrobacter_aerosaccus.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Pseudomonas.geniculata..EUPATH_0009257_Bacteria_teria_Xanthomonadales_Xanthomonadaceae_Xanthomonadaceae_unclassified_Pseudomonas_geniculata.}}{double COLUMN_DESCRIPTION}
  \item{\code{Peptoniphilus.sp..HMSC062D09..EUPATH_0009257_Bacteria_tes_Tissierellia_Tissierellales_Peptoniphilaceae_Peptoniphilus_Peptoniphilus_sp_HMSC062D09.}}{double COLUMN_DESCRIPTION}
  \item{\code{Staphylococcus.argenteus..EUPATH_0009257_Bacteria_teus.}}{double COLUMN_DESCRIPTION}
  \item{\code{Bifidobacterium.scardovii..EUPATH_0009257_Bacteria_tinobacteria_Bifidobacteriales_Bifidobacteriaceae_Bifidobacterium_Bifidobacterium_scardovii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.infantis..EUPATH_0009257_Bacteria_tis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Peptostreptococcus.anaerobius..EUPATH_0009257_Bacteria_tridia_Clostridiales_Peptostreptococcaceae_Peptostreptococcus_Peptostreptococcus_anaerobius.}}{double COLUMN_DESCRIPTION}
  \item{\code{Clostridium.paraputrificum..EUPATH_0009257_Bacteria_um.}}{double COLUMN_DESCRIPTION}
  \item{\code{Lactobacillus.crispatus..EUPATH_0009257_Bacteria_us.}}{double COLUMN_DESCRIPTION}
  \item{\code{Streptococcus.salivarius_CAG_79..EUPATH_0009257_Bacteria_utes_Bacilli_Lactobacillales_Streptococcaceae_Streptococcus_Streptococcus_salivarius_CAG_79.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Clostridium.clostridioforme..EUPATH_0009257_Bacteria_utes_Clostridia_Clostridiales_Lachnospiraceae_Lachnoclostridium_Clostridium_clostridioforme.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Clostridioides.difficile..EUPATH_0009257_Bacteria_utes_Clostridia_Clostridiales_Peptostreptococcaceae_Clostridioides_Clostridioides_difficile.}}{double COLUMN_DESCRIPTION}
  \item{\code{Agathobaculum.butyriciproducens..EUPATH_0009257_Bacteria_utes_Clostridia_Clostridiales_Ruminococcaceae_Agathobaculum_Agathobaculum_butyriciproducens.}}{double COLUMN_DESCRIPTION}
  \item{\code{Faecalibacterium.prausnitzii..EUPATH_0009257_Bacteria_utes_Clostridia_Clostridiales_Ruminococcaceae_Faecalibacterium_Faecalibacterium_prausnitzii.}}{double COLUMN_DESCRIPTION}
  \item{\code{Atopobium.minutum..EUPATH_0009257_Bacteria_utum.}}{double COLUMN_DESCRIPTION}
  \item{\code{Neisseria.subflava..EUPATH_0009257_Bacteria_va.}}{double COLUMN_DESCRIPTION}
  \item{\code{Veillonella.parvula..EUPATH_0009257_Bacteria_vula.}}{double COLUMN_DESCRIPTION}
  \item{\code{Anaerococcus.nagyae..EUPATH_0009257_Bacteria_yae.}}{double COLUMN_DESCRIPTION}
  \item{\code{Candida.dubliniensis..EUPATH_0009257_Eukaryota_Ascomycota_Saccharomycetes_Saccharomycetales_Debaryomycetaceae_Candida_Candida_dubliniensis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Candida.parapsilosis..EUPATH_0009257_Eukaryota_Ascomycota_Saccharomycetes_Saccharomycetales_Debaryomycetaceae_Candida_Candida_parapsilosis.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Malassezia.globosa..EUPATH_0009257_Eukaryota_a.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Malassezia.restricta..EUPATH_0009257_Eukaryota_asidiomycota_Malasseziomycetes_Malasseziales_Malasseziaceae_Malassezia_Malassezia_restricta.}}{logical COLUMN_DESCRIPTION}
  \item{\code{Candida.albicans..EUPATH_0009257_Eukaryota_cans.}}{double COLUMN_DESCRIPTION}
  \item{\code{Malassezia.sympodialis..EUPATH_0009257_Eukaryota_idiomycota_Malasseziomycetes_Malasseziales_Malasseziaceae_Malassezia_Malassezia_sympodialis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Malassezia.pachydermatis..EUPATH_0009257_Eukaryota_iomycota_Malasseziomycetes_Malasseziales_Malasseziaceae_Malassezia_Malassezia_pachydermatis.}}{double COLUMN_DESCRIPTION}
  \item{\code{Candida.tropicalis..EUPATH_0009257_Eukaryota_is.}}{logical COLUMN_DESCRIPTION}
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_84fcb69f4e/new/details}
}
\usage{
NEC_Metagenomic_data
}
\description{
This was a NICU study that conducted deep metagenomic profiling of samples collected longitudinally from 150 pre-term infants. 30 of these infants developed necrotizing enterocolitis. For this example data, we used the MicrobiomeDB to pull the relative abundance metagenomic tables for the 30 infants who developed NEC.
}
\keyword{datasets}
