% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEC_Repeated_data}
\alias{NEC_Repeated_data}
\title{NEC subset of Repeated Data from "Necrotizing enterocolitis is preceded by increased gut bacterial replication, Klebsiella, and fimbriae-encoding bacteria." Olm et al. Sci Adv 2019;5(12):eaax5727}
\format{
A data frame with 1097 rows and 4 variables:
\describe{
  \item{\code{Participant_repeated_measure_ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Participant_ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Days.of.period.NEC.diagnosed..days...EUPATH_0009213.}}{integer COLUMN_DESCRIPTION}
  \item{\code{Age..days...OBI_0001169.}}{integer COLUMN_DESCRIPTION}
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_84fcb69f4e/new/details}
}
\usage{
NEC_Repeated_data
}
\description{
This was a NICU study that conducted deep metagenomic profiling of samples collected longitudinally from 150 pre-term infants. 30 of these infants developed necrotizing enterocolitis. For this example data, we used the MicrobiomeDB to pull the relative abundance metagenomic tables at the species level.
}
\keyword{datasets}
