% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEC_Sample_data}
\alias{NEC_Sample_data}
\title{NEC subset of Sample Data from "Necrotizing enterocolitis is preceded by increased gut bacterial replication, Klebsiella, and fimbriae-encoding bacteria." Olm et al. Sci Adv 2019;5(12):eaax5727}
\format{
A data frame with 1124 rows and 24 variables:
\describe{
  \item{\code{Sample_ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Participant_repeated_measure_ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Participant_ID}}{character COLUMN_DESCRIPTION}
  \item{\code{Habitat..ENVO_01000739.}}{character COLUMN_DESCRIPTION}
  \item{\code{Host.body.habitat..UBERON_0000466.}}{character COLUMN_DESCRIPTION}
  \item{\code{Host.body.product..UBERON_0000463.}}{character COLUMN_DESCRIPTION}
  \item{\code{Host.body.site..UBERON_0000061.}}{character COLUMN_DESCRIPTION}
  \item{\code{Methicillin.resistant.Staphylococcus.aureus..ARO_3004306.}}{character COLUMN_DESCRIPTION}
  \item{\code{Sample.type..OBI_0100051.}}{character COLUMN_DESCRIPTION}
  \item{\code{Acyclovir..EUPATH_0009194.}}{character COLUMN_DESCRIPTION}
  \item{\code{Amoxicillin..EUPATH_0009195.}}{character COLUMN_DESCRIPTION}
  \item{\code{Ampicillin..EUPATH_0009196.}}{character COLUMN_DESCRIPTION}
  \item{\code{Cefazolin..EUPATH_0009197.}}{character COLUMN_DESCRIPTION}
  \item{\code{Cefepime..EUPATH_0009198.}}{character COLUMN_DESCRIPTION}
  \item{\code{Cefotaxime..EUPATH_0009199.}}{character COLUMN_DESCRIPTION}
  \item{\code{Claforan..EUPATH_0009200.}}{character COLUMN_DESCRIPTION}
  \item{\code{Clindamycin..EUPATH_0009201.}}{character COLUMN_DESCRIPTION}
  \item{\code{Fluconazole..EUPATH_0009202.}}{character COLUMN_DESCRIPTION}
  \item{\code{Gentamycin..EUPATH_0009203.}}{character COLUMN_DESCRIPTION}
  \item{\code{Nafcillin..EUPATH_0009204.}}{character COLUMN_DESCRIPTION}
  \item{\code{Nystatin..EUPATH_0009205.}}{character COLUMN_DESCRIPTION}
  \item{\code{Ofloxacin..EUPATH_0009206.}}{character COLUMN_DESCRIPTION}
  \item{\code{Vancomycin..EUPATH_0009207.}}{character COLUMN_DESCRIPTION}
  \item{\code{Zosyn..EUPATH_0009208.}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://microbiomedb.org/mbio/app/workspace/analyses/DS_84fcb69f4e/new/details}
}
\usage{
NEC_Sample_data
}
\description{
This was a NICU study that conducted deep metagenomic profiling of samples collected longitudinally from 150 pre-term infants. 30 of these infants developed necrotizing enterocolitis. For this example data, we used the MicrobiomeDB to pull the relative abundance metagenomic tables at the species level.
}
\keyword{datasets}
