% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phyloseq_to_Limon.R
\name{phyloseq_to_LIMON_Obj}
\alias{phyloseq_to_LIMON_Obj}
\title{Phyloseq to LIMON}
\usage{
phyloseq_to_LIMON_Obj(ps_object)
}
\arguments{
\item{ps_object}{a phyloseq object}
}
\value{
Returns a LIMON Object with a Count and SampleData table in the correct format
}
\description{
Create a formatted objects to run LIMON from a Phyloseq object
}
