% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIMON_DistrFit.R
\name{LIMON_DistrFit}
\alias{LIMON_DistrFit}
\title{Fit a zero inflated negative binomial linear mixed model with covariate correction}
\usage{
LIMON_DistrFit(
  Obj,
  Time,
  Subject,
  Covariates,
  model,
  distribution = "LMM",
  correlation = NULL
)
}
\arguments{
\item{Obj}{(Required) A LIMON Object}

\item{Time}{(Required) A column in "SampleData" for the Time variable. Must be numeric ordinal (1,2,3, etc)}

\item{Subject}{(Required) A column of subject IDs. Used as}

\item{Covariates}{(Required) Specified string of columns from SampleData to use as covariates in the model}

\item{model}{(Required) Specify the right side of the lmer equation (the covariates to correct by)}

\item{distribution}{(Optional) Specifies the distribution to fit for the linear mixed effect model. [Default] = "LMM", or the lmer() function from lme4. Other options include "LME.Z" = NBZIMM::lme.zig() , "GLMM.Z" = NBZIMM::lme.zig(), or "GLMM.NB" = NBZIMM::glmm.nb()}

\item{correlation}{(Optional) Option to specify the correlation structure for model fitting ([Default] of NULL specifies unstructured). Follows the same covariance structure names as in NBZIMM}
}
\value{
LIMON Object with the addition; (1) Model_data: contains the Counts and SampleData in Long Format for the model. (2) equation: the formula for mixed effect model. (3) Model_res: The results from running the linear mixed model. (4) Corrected_Counts: corrected count values with samples as rows and taxa as columns. (5) Corrected_Counts_Time: corrected count values per time point with samples as rows and taxa as columns.
}
\description{
This function fits a mixed effect model with user specified distribution, covariates and subjects as random effects. The returned value are counts per time point that have had the covariate effect removed from them. We recommend users examine the distribution of their data before selection the appropriate model distribution and covariance structure. The default fit is a Linear Mixed Effect model, but the other options include a Zero Inflated Gaussian, Zero Inflated Negative Binomial, and Negative Binomial linear mixed effect models. The function will first convert the count table and sample data into a long formatted model table. The model will be fit with the outcome variable as "asvcounts" and the predictors as covariates to correct by with subjects modeled as random effects. Corrected counts are calculated by subtracting the covariate estimated coefficient*covariate value from the observed counts. Please see the paper documentation for more information.
}
