% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIMON_IndNet.R
\name{LIMON_IndNet}
\alias{LIMON_IndNet}
\title{Individual Network Inferences}
\usage{
LIMON_IndNet(
  Obj,
  method,
  sel.criterion,
  lambda.min.ratio,
  nlambda,
  pulsar.select = TRUE,
  pulsar.params = list(),
  icov.select = pulsar.select,
  icov.select.params = pulsar.params,
  lambda.log = TRUE
)
}
\arguments{
\item{Obj}{LIMON Object from LIMON_NetInf_Time()}

\item{method}{Default "glasso". "mb" other option from SpiecEasi}

\item{sel.criterion}{Method for model selection. Either 'stars' [default] or 'bstars' (Bounded StARS)}

\item{lambda.min.ratio}{Scaling factor that will set the minimum lambda (sparsity) parameter}

\item{nlambda}{sparsity parameter (see SpiecEasi for more details)}

\item{pulsar.select}{Default TRUE}

\item{pulsar.params}{Default empty list()}

\item{icov.select}{Default "pulsar.select"}

\item{icov.select.params}{Default "pulsar.params"}

\item{lambda.log}{Default TRUE}
}
\value{
LIMON object with individual networks per sample stored in Individual_Networks by sample name
}
\description{
# Function to perform network inferences following method in lionessR - https://github.com/mararie/lionessR. Note: Consider using the same methods and parameters as in LIMON_NetInf_Time(). See the \pkg{SpiecEasi} documentation for more details.
}
