% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIMON_NetInf_Time.R
\name{LIMON_NetInf_Time}
\alias{LIMON_NetInf_Time}
\title{Infer Network per time point}
\usage{
LIMON_NetInf_Time(
  Obj,
  method,
  sel.criterion,
  lambda.min.ratio,
  nlambda,
  pulsar.select = TRUE,
  pulsar.params = list(),
  icov.select = pulsar.select,
  icov.select.params = pulsar.params,
  lambda.log = TRUE
)
}
\arguments{
\item{Obj}{LIMON Object after performing LIMON_DistrFit()}

\item{method}{Default "glasso". "mb" other option from SpiecEasi}

\item{sel.criterion}{Method for model selection. Either 'stars' [default] or 'bstars' (Bounded StARS)}

\item{lambda.min.ratio}{Scaling factor that will set the minimum lambda (sparsity) parameter}

\item{nlambda}{sparsity parameter (see SpiecEasi for more details)}

\item{pulsar.select}{Default TRUE}

\item{pulsar.params}{Default empty list()}

\item{icov.select}{Default "pulsar.select"}

\item{icov.select.params}{Default "pulsar.params"}

\item{lambda.log}{Default TRUE}
}
\value{
Returns a LIMON object with SpiecEasi networks and Covariance matrices by time added to it. SpiecEasi_Time_i: SpiecEasi object per time point. CovMatrix_Time_i: Covariance matrix per time point extracted from Spiec Easi using getOptCov(). CovMatrix_Diff_ji: Difference in covariance matrix between time j and time i.
}
\description{
Uses SPIEC-EASI to fit a network per Time point and return that network object, the optimal covariance matrix per time, and differences in covariance matrix per time. See the \pkg{SpiecEasi} documentation for more details.
}
